Fix memory leak in NativeSwing Auto-Shutdown thread
The NativeSwing Auto-Shutdown thread contains a memory leak because it contains an array that saves references to Thread objects.  When the application reduces the number of running threads, the Thread objects are not allowed to be garbage collected because the "activeThreads" array maintains references to them.  This patch fixes the issue by nulling out the thread references after they have been checked.